+++ include/piecewise-approximation/linear.hpp
@@ -6,25 +6,25 @@ namespace Swab {
     // Source path: src/piecewise-approximation/linear/swab.cpp
 
     struct Segment {
-        std::vector<long double> data;
+        std::vector<double> data;
         UpperHull l_cvx;
         LowerHull u_cvx;
 
-        Segment(std::vector<long double>& data, UpperHull& l_cvx, LowerHull& u_cvx);
+        Segment(std::vector<double>& data, UpperHull& l_cvx, LowerHull& u_cvx);
     };
 
     class Approximator {
         public:
-            static Line interpolate(std::vector<long double>& data);
-            static Line regression(int size, long double acc, long double avg_x, long double avg_y, long square);
-            static Line regression(std::vector<long double>& data);
-            static long double cal_error(std::vector<long double>& data, Line& line);
+            static Line interpolate(std::vector<double>& data);
+            static Line regression(int size, double acc, double avg_x, double avg_y, long square);
+            static Line regression(std::vector<double>& data);
+            static double cal_error(std::vector<double>& data, Line& line);
     };
 
     class Grouper {
         public:
             static void merge(Segment& s1, Segment& s2, std::string mode);
-            static long double merge_cost(Segment& s1, Segment& s2, std::string mode);
+            static double merge_cost(Segment& s1, Segment& s2, std::string mode);
             static bool bound_check(UpperHull& l_cvx, LowerHull& u_cvx, Line& line, int offset=0);
 
     };
@@ -32,20 +32,20 @@ namespace Swab {
     class Compression : public BaseCompression {
         private:
             int n_segment = 0;
-            long double error = 0;
+            double error = 0;
             std::string mode = "";
 
             bool first = true;
             UpperHull l_cvx;
             LowerHull u_cvx;
-            std::vector<long double> m_err;
-            std::vector<long double> window;
+            std::vector<double> m_err;
+            std::vector<double> window;
             std::vector<Segment> segments;
             
             // Parameters for incremental swab-regression
-            long double accumulate = 0;
-            long double average_x = 0;
-            long double average_y = 0;
+            double accumulate = 0;
+            double average_x = 0;
+            double average_y = 0;
             long accumulate_square = 0;
 
             void __bottom_up();
@@ -64,7 +64,7 @@ namespace Swab {
     class Decompression : public BaseDecompression {
         private:
             std::string mode = "";
-            long double pivot = INFINITY;
+            double pivot = INFINITY;
             
         protected:
             CSVObj* decompress(BinObj* compress_data) override;
@@ -82,10 +82,10 @@ namespace SwingFilter {
     
     class Compression : public BaseCompression {
         private:
-            long double error = 0;
+            double error = 0;
 
-            long double A_num = 0;
-            long double A_den = 0;
+            double A_num = 0;
+            double A_den = 0;
             Point2D* pivot = nullptr;
             Line* u_line = nullptr;
             Line* l_line = nullptr;
@@ -126,7 +126,7 @@ namespace SlideFilter {
     // Source path: src/piecewise-approximation/linear/slide-filter.cpp
     class Compression : public BaseCompression {
         private:
-            long double error = 0;
+            double error = 0;
             int segment_pos = 0;
             
             Point2D* pivot = nullptr;
@@ -142,7 +142,7 @@ namespace SlideFilter {
             Point2D* prev_end = nullptr;
 
             Line __fit();
-            long double __checkConnected();
+            double __checkConnected();
 
         protected:
             void compress(Univariate* data) override;
@@ -174,11 +174,11 @@ namespace CovariancePLA {
     // Source path: src/piecewise-approximation/linear/optimal-pla.cpp
     class Compression : public BaseCompression {
         private:
-            long double error = 0;
+            double error = 0;
 
-            long double accumulate = 0;
-            long double average_x = 0;
-            long double average_y = 0;
+            double accumulate = 0;
+            double average_x = 0;
+            double average_y = 0;
             long accumulate_square = 0;
             Line* line = nullptr;
             
@@ -215,7 +215,7 @@ namespace OptimalPLA {
     // Source path: src/piecewise-approximation/linear/optimal-pla.cpp
     class Compression : public BaseCompression {
         private:
-            long double error = 0;
+            double error = 0;
 
             Point2D* pivot = nullptr;
             Line* u_line = nullptr;
@@ -257,8 +257,8 @@ namespace ConnIPLA {
             LowerHull l_cvx;
             Point2D* pivot = nullptr;
             
-            void __optimal_approximate(Point2D& p, long double error);
-            void __fsw_approximate(Point2D& p, long double error);
+            void __optimal_approximate(Point2D& p, double error);
+            void __fsw_approximate(Point2D& p, double error);
 
         public:
             long t_start;
@@ -272,12 +272,12 @@ namespace ConnIPLA {
             ~LinearSegment();
 
             Line getLine();
-            void approximate(Point2D& p, long double error);
+            void approximate(Point2D& p, double error);
     };
     
     class Compression : public BaseCompression {
         private:
-            long double error = 0;
+            double error = 0;
 
             int phase = 0;
             LinearSegment* s_1 = nullptr;
@@ -317,10 +317,10 @@ namespace IOrientedPLA {
     // Source path: src/piecewise-approximation/linear/ioriented-pla.cpp
     class Compression : public BaseCompression {
         private:
-            long double scale = 1;
-            long double upshift = 0;
-            long double downshift = 0;
-            long double error = 0;
+            double scale = 1;
+            double upshift = 0;
+            double downshift = 0;
+            double error = 0;
 
             Point2D* pivot = nullptr;
             Line* u_line = nullptr;
@@ -345,9 +345,9 @@ namespace IOrientedPLA {
 
     class Decompression : public BaseDecompression {
         private:
-            long double scale = 1;
-            long double upshift = 0;
-            long double downshift = 0;
+            double scale = 1;
+            double upshift = 0;
+            double downshift = 0;
 
         protected:
             CSVObj* decompress(BinObj* compress_data) override;
@@ -390,18 +390,18 @@ namespace SemiOptimalPLA {
             ~OptimalPLA();
 
             void reconstructCvx();
-            std::pair<Point2D, bool> shrink(OptimalPLA* prev_seg, long double error);
-            std::pair<Point2D, bool> extendBackward(OptimalPLA* prev_seg, long double error);
+            std::pair<Point2D, bool> shrink(OptimalPLA* prev_seg, double error);
+            std::pair<Point2D, bool> extendBackward(OptimalPLA* prev_seg, double error);
             void updateExtrm();
             Point2D popLast(bool flag);
-            bool isSemiConnected(Line* line, long double bound);            
-            void approximate(Point2D& p, long double error);
+            bool isSemiConnected(Line* line, double bound);            
+            void approximate(Point2D& p, double error);
             void rconcate(std::vector<Point2D>& points);
     };
 
     class Compression : public BaseCompression {
         private:
-            long double error = 0;
+            double error = 0;
 
             OptimalPLA* seg_1 = nullptr;
             OptimalPLA* seg_2 = nullptr;
@@ -471,18 +471,18 @@ namespace SemiMixedPLA {
             ~OptimalPLA();
             
             void reconstructCvx();
-            std::pair<Point2D, bool> shrink(OptimalPLA* prev_seg, long double error);
-            std::pair<Point2D, bool> extendBackward(OptimalPLA* prev_seg, long double error);
+            std::pair<Point2D, bool> shrink(OptimalPLA* prev_seg, double error);
+            std::pair<Point2D, bool> extendBackward(OptimalPLA* prev_seg, double error);
             void updateExtrm();
             Point2D popLast(bool flag);
-            bool isSemiConnected(Line* line, long double bound);            
-            void approximate(Point2D& p, long double error);
+            bool isSemiConnected(Line* line, double bound);            
+            void approximate(Point2D& p, double error);
             void rconcate(std::vector<Point2D>& points);
     };
 
     class SemiOptimalPLA {
         private:
-            int __check(long double error); // check if two segments belong to which case
+            int __check(double error); // check if two segments belong to which case
 
         public:
             int seg_count = 0;
@@ -496,22 +496,22 @@ namespace SemiMixedPLA {
             SemiOptimalPLA(OptimalPLA* p_seg, Point2D* p_point);
             ~SemiOptimalPLA();
 
-            void finalize(long double error, long index);
+            void finalize(double error, long index);
             long getLastEnd();
             BinObj* getInter();
             OptimalPLA* getPendSeg();
             Point2D* getPendPoint();
-            void makeSemiConnect(int semi_case, long double error);
+            void makeSemiConnect(int semi_case, double error);
             // return 0: segments are not completed
             // return 1: case 1 of semi connected
             // return 2: case 2 of semi connected
             // return 3: case 3 of semi connected
-            int approximate(Point2D& p, long double error);
+            int approximate(Point2D& p, double error);
     };
 
     class Compression : public BaseCompression {
         private:
-            long double error = 0;
+            double error = 0;
 
             std::vector<Point2D> buffer;
             SemiOptimalPLA* semi_segs_1 = nullptr;
@@ -536,7 +536,7 @@ namespace SemiMixedPLA {
 
     class Decompression : public BaseDecompression {
         private:
-            const long double EPSILON = 1e-6;
+            const double EPSILON = 1e-6;
             
             long index = 0;
             Point2D* prev_end = nullptr;
@@ -556,80 +556,80 @@ namespace MixPiece {
     // Source path: src/piecewise-approximation/linear/mix-piece.cpp
     struct Segment {
         long t;
-        long double aMin;
-        long double aMax;
-        long double b;
-        long double a;
+        double aMin;
+        double aMax;
+        double b;
+        double a;
 
-        Segment(long t, long double a, long double b);
-        Segment(long t, long double aMin, long double aMax, long double b);
+        Segment(long t, double a, double b);
+        Segment(long t, double aMin, double aMax, double b);
     };
 
     struct BBlock {
 
         struct Block {
-            long double a_u;
-            long double a_l;
+            double a_u;
+            double a_l;
             std::vector<long> t;
 
             Block();
-            Block(long double a_u, long double a_l, long t);
+            Block(double a_u, double a_l, long t);
         };
 
-        long double b;
+        double b;
         std::vector<Block> blocks;
 
-        BBlock(long double b);
-        BBlock(long double b, long double a_u, long double a_l, long t);
+        BBlock(double b);
+        BBlock(double b, double a_u, double a_l, long t);
 
         void sort();
         Segment pop_back();
-        void push_back(long double a_u, long double a_l, long t);
-        bool is_intersect(long double a_u, long double a_l);
-        void intersect(long double a_u, long double a_l, long t);
+        void push_back(double a_u, double a_l, long t);
+        bool is_intersect(double a_u, double a_l);
+        void intersect(double a_u, double a_l, long t);
     };
 
     struct ABlock {
 
         struct Block {
-            long double b;
+            double b;
             long t;
 
-            Block(long double b, long t);
+            Block(double b, long t);
         };
 
-        long double a_u;
-        long double a_l;
+        double a_u;
+        double a_l;
         std::vector<Block> blocks;
 
         ABlock();
-        ABlock(long double b, long double a_u, long double a_l, long t);
+        ABlock(double b, double a_u, double a_l, long t);
 
         void sort();
-        bool is_intersect(long double a_u, long double a_l);
-        void intersect(long double b, long double a_u, long double a_l, long t);
+        bool is_intersect(double a_u, double a_l);
+        void intersect(double b, double a_u, double a_l, long t);
     };
 
     class Compression : public BaseCompression {
         private:
-            long double error = 0;
+            double error = 0;
             int n_segment = 0;
 
             int flag = 0;
             bool floor_flag = true;
             bool ceil_flag = true;
-            long double slp_u_1 = INFINITY;
-            long double slp_u_2 = INFINITY;
-            long double slp_l_1 = -INFINITY;
-            long double slp_l_2 = -INFINITY;
+            double slp_u_1 = INFINITY;
+            double slp_u_2 = INFINITY;
+            double slp_l_1 = -INFINITY;
+            double slp_l_2 = -INFINITY;
             long t_s = 0;
-            long double b_1 = 0;
-            long double b_2 = 0;
+            double b_1 = 0;
+            double b_2 = 0;
 
             std::vector<BBlock> BBlocks;
             std::vector<ABlock> ABlocks;
             std::vector<Segment> r_blocks;
-            std::map<long double, std::vector<Segment>> intervals;
+            std::map<double, std::vector<Segment>> intervals;
 
             long index = 0;
             int curr_segment = 0;
